+++ src/pages/Home/components/NewCycleForm/index.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react'
 import { useFormContext } from 'react-hook-form'
 
-import { CyclesContext } from '../..'
+import { CyclesContext } from '../../../../contexts/CyclesContext'
 
 import { FormContainer, MinutesAmountInput, TaskInput } from './styles'
